Split EditCaseInfoForm into 4 separate forms
