+++ src/containers/participant/ParticipantProfileContainer.js
@@ -9,7 +9,7 @@ import ParticipantProfile from './ParticipantProfile';
 import PrintWorkScheduleContainer from './print/PrintWorkScheduleContainer';
 import EditPersonAndContactsForm from './EditPersonAndContactsForm';
 import EditEnrollmentDatesForm from './EditEnrollmentDatesForm';
-import EditCaseInfoForm from './EditCaseInfoForm';
+import EditCaseInfoForm from './cases/EditCaseInfoForm';
 import * as Routes from '../../core/router/Routes';
 
 import { PERSON, STATE } from '../../utils/constants/ReduxStateConsts';
